delet post

+++ server/src/user/use_case/deletePost.js
@@ -5,7 +5,9 @@ module.exports = (repository) => {
     return new Promise((resolve, reject) => {
       try {
         repository.delete(id).then(async (post) => {
-          await deleteFile(post.image);
+          if (post.image) {
+            await deleteFile(post.image);
+          }
           repository.delete(id);
           resolve({
             msg: "post deleted",
